+++ crikit/ui/qt_PlotEffect_KK.py
@@ -1,17 +1,19 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file '.\qt5-creator\ui_PlotEffect_KK.ui'
+# Form implementation generated from reading ui file '.\ui_PlotEffect_KK.ui'
 #
-# Created by: PyQt5 UI code generator 5.5.1
+# Created by: PyQt5 UI code generator 5.13.0
 #
 # WARNING! All changes made in this file will be lost!
 
+
 from PyQt5 import QtCore, QtGui, QtWidgets
+
 
 class Ui_Form(object):
     def setupUi(self, Form):
         Form.setObjectName("Form")
-        Form.resize(752, 166)
+        Form.resize(752, 215)
         Form.setStyleSheet("font: 10pt \"Arial\";")
         self.horizontalLayout = QtWidgets.QHBoxLayout(Form)
         self.horizontalLayout.setObjectName("horizontalLayout")
@@ -65,6 +67,9 @@ class Ui_Form(object):
         self.label_4 = QtWidgets.QLabel(self.frame)
         self.label_4.setObjectName("label_4")
         self.verticalLayout_5.addWidget(self.label_4, 0, QtCore.Qt.AlignHCenter|QtCore.Qt.AlignTop)
+        self.checkBoxConjugate = QtWidgets.QCheckBox(self.frame)
+        self.checkBoxConjugate.setObjectName("checkBoxConjugate")
+        self.verticalLayout_5.addWidget(self.checkBoxConjugate)
         self.checkBoxNRBNorm = QtWidgets.QCheckBox(self.frame)
         self.checkBoxNRBNorm.setChecked(True)
         self.checkBoxNRBNorm.setObjectName("checkBoxNRBNorm")
@@ -86,6 +91,21 @@ class Ui_Form(object):
         self.spinBoxPadFactor.setObjectName("spinBoxPadFactor")
         self.horizontalLayout_2.addWidget(self.spinBoxPadFactor)
         self.verticalLayout_5.addLayout(self.horizontalLayout_2)
+        self.horizontalLayout_3 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_3.setContentsMargins(-1, 0, -1, -1)
+        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
+        self.label_2 = QtWidgets.QLabel(self.frame)
+        self.label_2.setObjectName("label_2")
+        self.horizontalLayout_3.addWidget(self.label_2)
+        self.spinBoxEdge = QtWidgets.QSpinBox(self.frame)
+        self.spinBoxEdge.setMinimum(1)
+        self.spinBoxEdge.setMaximum(800)
+        self.spinBoxEdge.setObjectName("spinBoxEdge")
+        self.horizontalLayout_3.addWidget(self.spinBoxEdge)
+        self.verticalLayout_5.addLayout(self.horizontalLayout_3)
+        self.checkBoxShowReal = QtWidgets.QCheckBox(self.frame)
+        self.checkBoxShowReal.setObjectName("checkBoxShowReal")
+        self.verticalLayout_5.addWidget(self.checkBoxShowReal)
         spacerItem3 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
         self.verticalLayout_5.addItem(spacerItem3)
         self.horizontalLayout.addWidget(self.frame)
@@ -100,7 +120,9 @@ class Ui_Form(object):
         self.label_9.setText(_translate("Form", "NRB Bias"))
         self.label.setText(_translate("Form", "Phase Bias"))
         self.label_4.setText(_translate("Form", "Options"))
+        self.checkBoxConjugate.setText(_translate("Form", "Conjugate"))
         self.checkBoxNRBNorm.setText(_translate("Form", "Norm to NRB"))
         self.checkBoxLockBias.setText(_translate("Form", "Lock NRB-to-CARS Bias"))
         self.label_5.setText(_translate("Form", "Pad Factor"))
-
+        self.label_2.setText(_translate("Form", "N Edge Pix (Avg)"))
+        self.checkBoxShowReal.setText(_translate("Form", "Show Real Part"))