+++ crikit/ui/widget_KK.py
@@ -71,8 +71,8 @@ class widgetKK(_AbstractPlotEffectPlugin):
                       }
                       
     def __init__(self, cars_amp_offset=0.0, nrb_amp_offset=0.0,
-                 phase_offset=0.0, norm_to_nrb=True, pad_factor=1, 
-                 parent = None):
+                 conjugate=False, phase_offset=0.0, norm_to_nrb=True, pad_factor=1,
+                 n_edge=30, parent=None):
         
         super(widgetKK, self).__init__(parent)
         
@@ -80,33 +80,43 @@ class widgetKK(_AbstractPlotEffectPlugin):
         self.ui.setupUi(self)
         
         self.lock_cars_nrb_bias = True
-        
+        self.show_real = False
+
         # Update parameter dict
         self.parameters['cars_amp_offset'] = cars_amp_offset
         self.parameters['nrb_amp_offset'] = nrb_amp_offset
         self.parameters['phase_offset'] = phase_offset
         self.parameters['norm_to_nrb'] = norm_to_nrb
         self.parameters['pad_factor'] = pad_factor
+        self.parameters['n_edge'] = n_edge
+        self.parameters['conjugate'] = conjugate
                 
         self.setupKK()
         
     def setupKK(self):
 
         self.ui.checkBoxNRBNorm.setChecked(self.parameters['norm_to_nrb'])
+        self.ui.checkBoxConjugate.setChecked(self.parameters['conjugate'])
         self.ui.checkBoxLockBias.setChecked(self.lock_cars_nrb_bias)
-        
+        self.ui.checkBoxShowReal.setChecked(self.show_real)
+
         self.ui.spinBoxCARSBias.setValue(self.parameters['cars_amp_offset'])
         self.ui.spinBoxNRBBias.setValue(self.parameters['nrb_amp_offset'])
         self.ui.spinBoxPhaseLin.setValue(self.parameters['phase_offset'])
         self.ui.spinBoxPadFactor.setValue(self.parameters['pad_factor'])
+        self.ui.spinBoxEdge.setValue(self.parameters['n_edge'])
 
         # SIGNALS & SLOTS
         self.ui.spinBoxCARSBias.editingFinished.connect(self.spinBoxChanged)
         self.ui.spinBoxNRBBias.editingFinished.connect(self.spinBoxChanged)
         self.ui.spinBoxPhaseLin.editingFinished.connect(self.spinBoxChanged)
         self.ui.spinBoxPadFactor.editingFinished.connect(self.spinBoxChanged)
+        self.ui.spinBoxEdge.editingFinished.connect(self.spinBoxChanged)
         
         self.ui.checkBoxNRBNorm.clicked.connect(self.changeCheckBoxNRBNorm)
+        self.ui.checkBoxConjugate.clicked.connect(self.changeCheckBoxConjugate)
+        self.ui.checkBoxShowReal.clicked.connect(self.changeShowReal)
+
         self.ui.checkBoxLockBias.clicked.connect(self.changeCheckBoxLockBias)
 
         self.ui.spinBoxNRBBias.setEnabled(not self.lock_cars_nrb_bias)
@@ -130,19 +140,26 @@ class widgetKK(_AbstractPlotEffectPlugin):
                
         cars_amp_offset = self.parameters['cars_amp_offset']
         nrb_amp_offset = self.parameters['nrb_amp_offset'] 
+        conjugate = self.parameters['conjugate']
         phase_offset = self.parameters['phase_offset'] 
         norm_to_nrb = self.parameters['norm_to_nrb']
         pad_factor = self.parameters['pad_factor']
+        n_edge = self.parameters['n_edge']
         
         _kk = _KramersKronig(cars_amp_offset=cars_amp_offset,
                              nrb_amp_offset=nrb_amp_offset,
+                             conjugate=conjugate,
                              phase_offset=phase_offset,
                              norm_to_nrb=norm_to_nrb,
-                             pad_factor=pad_factor)
+                             pad_factor=pad_factor,
+                             n_edge=n_edge)
 
         data_out = _kk.calculate(cars, nrb)
         
-        return data_out.imag
+        if self.show_real:
+            return data_out.real
+        else:
+            return data_out.imag
 
     def spinBoxChanged(self):
         """
@@ -166,7 +183,8 @@ class widgetKK(_AbstractPlotEffectPlugin):
             
         elif sdr == self.ui.spinBoxPadFactor:
             self.parameters['pad_factor'] = self.ui.spinBoxPadFactor.value()
-            
+        elif sdr == self.ui.spinBoxEdge:
+            self.parameters['n_edge'] = self.ui.spinBoxEdge.value()
         self.changed.emit()
 
     def changeCheckBoxLockBias(self):
@@ -185,10 +203,19 @@ class widgetKK(_AbstractPlotEffectPlugin):
             self.parameters['norm_to_nrb'] = False
         self.changed.emit()
 
-    def changeSpinBoxPadFactor(self):
-        self.parameters['pad_factor'] = self.ui.spinBoxPadFactor.value()
+    def changeCheckBoxConjugate(self):
+        if self.ui.checkBoxConjugate.isChecked():
+            self.parameters['conjugate'] = True
+        else:
+            self.parameters['conjugate'] = False
         self.changed.emit()
         
+    def changeShowReal(self):
+        if self.show_real:
+            self.show_real = False
+        else:
+            self.show_real = True
+        self.changed.emit()
         
 if __name__ == '__main__':
     import sys as _sys